Declare binary integer constants using binary prefix instead of using special macro
GCC has supported binary prefix for integer constants as an extension
(adapted from AVR-GCC local patches) since release 4.3.0 in early 2008.

Now, since this binary prefix is also in C++14 and will probably be
included in the next C standard it is going to stay, so let's use it
instead of hand-crafted BIN2BYTE() macro for our binary integer constants.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>