Avoid build error on binutils without "avr" size format patch
"avr" format for avr-size (also available via "-C" option) is provided by
a local patch included in some distributions of AVR binutils.
That patch was rejected by binutils upstream, so this format is unlikely to
be supported by future stock binutils releases.

Let's use a standard sysv format for printing code and data size in case
"avr" format is not available to avoid ending an otherwise successful
build with an error.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>