Fix EEPROM reading / writing at nonzero base address
According to STK500 protocol spec only word-addressed memories (program
flash) have addresses specified in word units in "load address" command,
however the bootloader had always multiplied address received from this
command by 2.

This resulted in a wrong base address for EEPROM reading / writing when
the requested base address was nonzero.
Since maximum message size is 275 bytes all EEPROM reads / writes to
address that were slightly less than this number and higher were being
directed to wrong cells.

This bug fix was the sole change in 2016.07.24 release by Peter Fleury.

Use this opportunity to also make sure that bit 31 is unset in the received
address - it is used as a request to execute a "load extended address"
command on a target by hardware programmer, so it has no meaning for a
bootloader.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>