Allow having separate build tree(s) from the source one
This commit adds an ability to have separate build tree(s) from the source
one, useful for keeping bootloaders for different boards in different
directories (for example their project ones).

Current build directory is selected by providing a "DESTDIR" parameter to
make.
By default (when the "DESTDIR" parameter is missing), the source directory
is used as the build directory, so there is no change in behavior in this
case.

Only the Makefile and documentation are modified by this change, the actual
source code is not touched.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>